+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog
 0.8.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Fix to normalize brackets in package [extras]
+  [datakurre]
 
 0.8.0 (2015-07-10)
 ------------------